Cleanup AWS Bedrock completions provider
